backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
redis==5.1.1
python-dotenv==1.0.1
aiohttp==3.10.10
cloudscraper==1.2.71
brotli==1.1.0

